full system working right now
