+++ src/Mementic_frontend/src/hooks/useMarketplaceFilters.js
@@ -1,0 +1,2 @@
+// This hook is no longer needed as filtering logic is moved back to the main component
+// Keeping file for potential future use
